requirements.txt
# Google API Dependencies
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
google-auth>=2.23.0

# MCP Framework
mcp>=0.9.0

# Async and HTTP
aiohttp>=3.9.0
asyncio>=3.4.3

# Data Validation
pydantic>=2.5.0

# Utilities
python-dotenv>=1.0.0
cryptography>=41.0.0
cachetools>=5.3.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
